+++ frontend/src/components/subject/modify/header/ModifySubjectHeaderCard.vue
