+++ frontend/src/components/subject/modify/header/ModifySubjectHeaderImage.vue
