--- frontend/src/components/subject/modify/patch/PatchSubjectBody.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-
-</template>
-
-<style scoped>
-
-</style>
