+++ frontend/src/router/index.ts
@@ -84,12 +84,12 @@ const router = createRouter({
         {
             path: "/subjects/create",
             name: "create-subject",
-            component: () => import("../views/subject/CreateSubjectView.vue"),
+            component: () => import("../views/subject/modify/CreateSubjectView.vue"),
         },
         {
           path: "/subjects/:subjectId(\\d+)/edit",
             name: "edit-subject",
-            component: () => import("../views/subject/PatchSubjectView.vue"),
+            component: () => import("../views/subject/modify/PatchSubjectView.vue"),
             props: (route) => ({ subjectId: Number(route.params.subjectId) }),
         },
         {
