+++ frontend/src/views/subject/modify/CreateSubjectView.vue
@@ -32,7 +32,7 @@
                     </v-card>
                 </v-dialog>
                 <div class="flex-container">
-                    <CreateSubjectHeaderContainer
+                    <ModifySubjectHeaderContainer
                         :title="$t('create_subject.new_subject')"
                         :image-path="`https://www.ugent.be/img/dcom/faciliteiten/ufo-logo.png`"
                         :subject-name="subjectName"
@@ -46,14 +46,14 @@
                         @update:subject-mail="onSubjectMailUpdated"
                         @update:current-user-as-instructor="currentUserAsInstructor = $event"
                     >
-                    </CreateSubjectHeaderContainer>
-                    <CreateSubjectBody
+                    </ModifySubjectHeaderContainer>
+                    <ModifySubjectBody
                         :current-user="currentUser"
                         :instructors="shownInstructors"
                         @add-instructor="addInstructor"
                         @remove-instructor="removeInstructor"
                     >
-                    </CreateSubjectBody>
+                    </ModifySubjectBody>
                 </div>
 
                 <div class="confirm-btn-container">
@@ -76,8 +76,8 @@ import { useCreateSubjectInstructorMutation, useCreateSubjectMutation } from "@/
 import type SubjectForm from "@/models/Subject";
 import type User from "@/models/User";
 import { useCurrentUserQuery } from "@/queries/User";
-import CreateSubjectHeaderContainer from "@/components/subject/modify/create/header/CreateSubjectHeaderContainer.vue";
-import CreateSubjectBody from "@/components/subject/modify/create/body/CreateSubjectBody.vue";
+import ModifySubjectHeaderContainer from "@/components/subject/modify/header/ModifySubjectHeaderContainer.vue";
+import ModifySubjectBody from "@/components/subject/modify/body/ModifySubjectBody.vue";
 import { useRouter } from "vue-router";
 
 const snackbar = ref(false);
